requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Pillow==10.2.0
bleach==6.1.0
Werkzeug==3.0.1
SQLAlchemy==2.0.28
mysqlclient==2.1.0 